// packages/storefront/src/app/common/generated-types.ts
export type Maybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};

export type CustomField = {
  description?: Maybe<Array<LocalizedString>>;
  internal?: Maybe<Scalars['Boolean']>;
  label?: Maybe<Array<LocalizedString>>;
  list: Scalars['Boolean'];
  name: Scalars['String'];
  readonly?: Maybe<Scalars['Boolean']>;
  type: Scalars['String'];
};

export type ErrorResult = {
  errorCode: ErrorCode;
  message: Scalars['String'];
};

export type Node = {
  id: Scalars['ID'];
};
// ...
export type Asset = Node & {
  __typename?: 'Asset';
  createdAt: Scalars['DateTime'];
  customFields?: Maybe<Scalars['JSON']>;
  fileSize: Scalars['Int'];
  focalPoint?: Maybe<Coordinate>;
  height: Scalars['Int'];
  id: Scalars['ID'];
  mimeType: Scalars['String'];
  name: Scalars['String'];
  preview: Scalars['String'];
  source: Scalars['String'];
  type: AssetType;
  updatedAt: Scalars['DateTime'];
  width: Scalars['Int'];
};
// ...
export type ConfigurableOperation = {
  __typename?: 'ConfigurableOperation';
  args: Array<ConfigArg>;
  code: Scalars['String'];
};
// ...
export type Discount = {
  __typename?: 'Discount';
  adjustmentSource: Scalars['String'];
  amount: Scalars['Int'];
  amountWithTax: Scalars['Int'];
  description: Scalars['String'];
  type: AdjustmentType;
};
// ...
export type Order = Node & {
  __typename?: 'Order';
  /** An order is active as long as the payment process has not been completed */
  active: Scalars['Boolean'];
  /**
   * Order-level adjustments to the order total, such as discounts from promotions
   * @deprecated Use `discounts` instead
   */
  adjustments: Array<Adjustment>;
  billingAddress?: Maybe<OrderAddress>;
  /** A unique code for the Order */
  code: Scalars['String'];
  /** An array of all coupon codes applied to the Order */
  couponCodes: Array<Scalars['String']>;
  createdAt: Scalars['DateTime'];
  currencyCode: CurrencyCode;
  customFields?: Maybe<Scalars['JSON']>;
  customer?: Maybe<Customer>;
  discounts: Array<Discount>;
  fulfillments?: Maybe<Array<Fulfillment>>;
  history: HistoryEntryList;
  id: Scalars['ID'];
  lines: Array<OrderLine>;
  /**
   * The date & time that the Order was placed, i.e. the Customer
   * completed the checkout and the Order is no longer "active"
   */
  orderPlacedAt?: Maybe<Scalars['DateTime']>;
  payments?: Maybe<Array<Payment>>;
  /** Promotions applied to the order. Only gets populated after the payment process has completed. */
  promotions: Array<Promotion>;
  shipping: Scalars['Int'];
  shippingAddress?: Maybe<OrderAddress>;
  shippingLines: Array<ShippingLine>;
  shippingWithTax: Scalars['Int'];
  state: Scalars['String'];
  /**
   * The subTotal is the total of all OrderLines in the Order. This figure also includes any Order-level
   * discounts which have been prorated (proportionally distributed) amongst the OrderItems.
   * To get a total of all OrderLines which does not account for prorated discounts, use the
   * sum of `OrderLine.discountedLinePrice` values.
   */
  subTotal: Scalars['Int'];
  /** Same as subTotal, but inclusive of tax */
  subTotalWithTax: Scalars['Int'];
  /**
   * Surcharges are arbitrary modifications to the Order total which are neither
   * ProductVariants nor discounts resulting from applied Promotions. For example,
   * one-off discounts based on customer interaction, or surcharges based on payment
   * methods.
   */
  surcharges: Array<Surcharge>;
  /** A summary of the taxes being applied to this Order */
  taxSummary: Array<OrderTaxSummary>;
  /** Equal to subTotal plus shipping */
  total: Scalars['Int'];
  totalQuantity: Scalars['Int'];
  /** The final payable amount. Equal to subTotalWithTax plus shippingWithTax */
  totalWithTax: Scalars['Int'];
  updatedAt: Scalars['DateTime'];
};
// ...
export type OrderLine = Node & {
  __typename?: 'OrderLine';
  /** @deprecated Use `discounts` instead */
  adjustments: Array<Adjustment>;
  createdAt: Scalars['DateTime'];
  customFields?: Maybe<Scalars['JSON']>;
  /** The price of the line including discounts, excluding tax */
  discountedLinePrice: Scalars['Int'];
  /** The price of the line including discounts and tax */
  discountedLinePriceWithTax: Scalars['Int'];
  /**
   * The price of a single unit including discounts, excluding tax.
   *
   * If Order-level discounts have been applied, this will not be the
   * actual taxable unit price (see `proratedUnitPrice`), but is generally the
   * correct price to display to customers to avoid confusion
   * about the internal handling of distributed Order-level discounts.
   */
  discountedUnitPrice: Scalars['Int'];
  /** The price of a single unit including discounts and tax */
  discountedUnitPriceWithTax: Scalars['Int'];
  discounts: Array<Discount>;
  featuredAsset?: Maybe<Asset>;
  id: Scalars['ID'];
  items: Array<OrderItem>;
  /** The total price of the line excluding tax and discounts. */
  linePrice: Scalars['Int'];
  /** The total price of the line including tax bit excluding discounts. */
  linePriceWithTax: Scalars['Int'];
  /** The total tax on this line */
  lineTax: Scalars['Int'];
  order: Order;
  productVariant: ProductVariant;
  /**
   * The actual line price, taking into account both item discounts _and_ prorated (proportially-distributed)
   * Order-level discounts. This value is the true economic value of the OrderLine, and is used in tax
   * and refund calculations.
   */
  proratedLinePrice: Scalars['Int'];
  /** The proratedLinePrice including tax */
  proratedLinePriceWithTax: Scalars['Int'];
  /**
   * The actual unit price, taking into account both item discounts _and_ prorated (proportially-distributed)
   * Order-level discounts. This value is the true economic value of the OrderItem, and is used in tax
   * and refund calculations.
   */
  proratedUnitPrice: Scalars['Int'];
  /** The proratedUnitPrice including tax */
  proratedUnitPriceWithTax: Scalars['Int'];
  quantity: Scalars['Int'];
  taxLines: Array<TaxLine>;
  taxRate: Scalars['Float'];
  /** @deprecated Use `linePriceWithTax` instead */
  totalPrice: Scalars['Int'];
  /** The price of a single unit, excluding tax and discounts */
  unitPrice: Scalars['Int'];
  /** Non-zero if the unitPrice has changed since it was initially added to Order */
  unitPriceChangeSinceAdded: Scalars['Int'];
  /** The price of a single unit, including tax but excluding discounts */
  unitPriceWithTax: Scalars['Int'];
  /** Non-zero if the unitPriceWithTax has changed since it was initially added to Order */
  unitPriceWithTaxChangeSinceAdded: Scalars['Int'];
  updatedAt: Scalars['DateTime'];
};
// ...
export type ProductVariant = Node & {
  __typename?: 'ProductVariant';
  assets: Array<Asset>;
  createdAt: Scalars['DateTime'];
  currencyCode: CurrencyCode;
  customFields?: Maybe<Scalars['JSON']>;
  facetValues: Array<FacetValue>;
  featuredAsset?: Maybe<Asset>;
  id: Scalars['ID'];
  languageCode: LanguageCode;
  name: Scalars['String'];
  options: Array<ProductOption>;
  price: Scalars['Int'];
  /** @deprecated price now always excludes tax */
  priceIncludesTax: Scalars['Boolean'];
  priceWithTax: Scalars['Int'];
  product: Product;
  productId: Scalars['ID'];
  sku: Scalars['String'];
  stockLevel: Scalars['String'];
  taxCategory: TaxCategory;
  taxRateApplied: TaxRate;
  translations: Array<ProductVariantTranslation>;
  updatedAt: Scalars['DateTime'];
};
// ...
export type ShippingLine = {
  __typename?: 'ShippingLine';
  discountedPrice: Scalars['Int'];
  discountedPriceWithTax: Scalars['Int'];
  discounts: Array<Discount>;
  price: Scalars['Int'];
  priceWithTax: Scalars['Int'];
  shippingMethod: ShippingMethod;
};

export type ShippingMethod = Node & {
  __typename?: 'ShippingMethod';
  calculator: ConfigurableOperation;
  checker: ConfigurableOperation;
  code: Scalars['String'];
  createdAt: Scalars['DateTime'];
  customFields?: Maybe<ShippingMethodCustomFields>;
  description: Scalars['String'];
  fulfillmentHandlerCode: Scalars['String'];
  id: Scalars['ID'];
  name: Scalars['String'];
  translations: Array<ShippingMethodTranslation>;
  updatedAt: Scalars['DateTime'];
};
// ...
  export type Variables = UpdateAddressMutationVariables;
// ...
  export type ActiveOrder = NonNullable<GetShippingAddressQuery['activeOrder']>;
// ...
  export type FeaturedAsset = NonNullable<
    NonNullable<
      NonNullable<NonNullable<GetCollectionsQuery['collections']>['items']>[number]
    >['featuredAsset']
  >;
// ...
export namespace Asset {
  export type Fragment = AssetFragment;
  export type FocalPoint = NonNullable<AssetFragment['focalPoint']>;
}

export namespace Cart {
  export type Fragment = CartFragment;
  export type Lines = NonNullable<NonNullable<CartFragment['lines']>[number]>;
  export type FeaturedAsset = NonNullable<
    NonNullable<NonNullable<CartFragment['lines']>[number]>['featuredAsset']
  >;
  export type ProductVariant = NonNullable<
    NonNullable<NonNullable<CartFragment['lines']>[number]>['productVariant']
  >;
  export type Discounts = NonNullable<
    NonNullable<NonNullable<NonNullable<CartFragment['lines']>[number]>['discounts']>[number]
  >;
  export type ShippingLines = NonNullable<NonNullable<CartFragment['shippingLines']>[number]>;
  export type ShippingMethod = NonNullable<
    NonNullable<NonNullable<CartFragment['shippingLines']>[number]>['shippingMethod']
  >;
  export type _Discounts = NonNullable<NonNullable<CartFragment['discounts']>[number]>;
}

export namespace Country {
  export type Fragment = CountryFragment;
}

export namespace OrderAddress {
  export type Fragment = OrderAddressFragment;
}

export namespace Address {
  export type Fragment = AddressFragment;
  export type Country = NonNullable<AddressFragment['country']>;
}

export namespace ErrorResult {
  export type Fragment = ErrorResultFragment;
}

export namespace GetActiveOrder {
  export type Variables = GetActiveOrderQueryVariables;
  export type Query = GetActiveOrderQuery;
  export type ActiveOrder = NonNullable<GetActiveOrderQuery['activeOrder']>;
}
// ...
export type GetCollectionsQuery = { __typename?: 'Query' } & {
  collections: { __typename?: 'CollectionList' } & {
    items: Array<
      { __typename?: 'Collection' } & Pick<Collection, 'id' | 'name' | 'slug'> & {
          parent?: Maybe<{ __typename?: 'Collection' } & Pick<Collection, 'id' | 'name' | 'slug'>>;
          featuredAsset?: Maybe<{ __typename?: 'Asset' } & AssetFragment>;
        }
    >;
  };
};

export type AssetFragment = { __typename?: 'Asset' } & Pick<
  Asset,
  'id' | 'width' | 'height' | 'name' | 'preview'
> & { focalPoint?: Maybe<{ __typename?: 'Coordinate' } & Pick<Coordinate, 'x' | 'y'>> };

export type CartFragment = { __typename?: 'Order' } & Pick<
  Order,
  | 'id'
  | 'code'
  | 'state'
  | 'active'
  | 'totalQuantity'
  | 'subTotal'
  | 'subTotalWithTax'
  | 'total'
  | 'totalWithTax'
  | 'shipping'
  | 'shippingWithTax'
> & {
    lines: Array<
      { __typename?: 'OrderLine' } & Pick<
        OrderLine,
        | 'id'
        | 'unitPrice'
        | 'unitPriceWithTax'
        | 'quantity'
        | 'linePriceWithTax'
        | 'discountedLinePriceWithTax'
      > & {
          featuredAsset?: Maybe<{ __typename?: 'Asset' } & AssetFragment>;
          productVariant: { __typename?: 'ProductVariant' } & Pick<ProductVariant, 'id' | 'name'>;
          discounts: Array<
            { __typename?: 'Discount' } & Pick<
              Discount,
              'amount' | 'amountWithTax' | 'description' | 'adjustmentSource' | 'type'
            >
          >;
        }
    >;
    shippingLines: Array<
      { __typename?: 'ShippingLine' } & Pick<ShippingLine, 'priceWithTax'> & {
          shippingMethod: { __typename?: 'ShippingMethod' } & Pick<
            ShippingMethod,
            'id' | 'code' | 'name' | 'description'
          >;
        }
    >;
    discounts: Array<
      { __typename?: 'Discount' } & Pick<
        Discount,
        'amount' | 'amountWithTax' | 'description' | 'adjustmentSource' | 'type'
      >
    >;
  };
// ...
export type GetActiveOrderQueryVariables = Exact<{ [key: string]: never }>;

export type GetActiveOrderQuery = { __typename?: 'Query' } & {
  activeOrder?: Maybe<{ __typename?: 'Order' } & CartFragment>;
};